js/xpconnect/tests/mochitest/test_shadowRealm.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test for ShadowRealms</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
  <iframe id="ifr"></iframe>
</head>

<body>
  <p id="display"></p>
  <script type="application/javascript">
    SimpleTest.waitForExplicitFinish();
    info("running")

    let realm = new ShadowRealm();

    let install = (fun, internal_name) => {
        let installer = realm.evaluate(`var ${internal_name}; (x) => { ${internal_name} = x}`);
        installer(fun);
    }

    install(info, "log");
    install(is, "is");
    realm.evaluate(`is(true, true, 'inside realm')`);

    is(realm.evaluate("10"), 10, "ten is ten");

    SimpleTest.finish();
  </script>
</body>

</html>
